+++ R/normalize_priors.R
@@ -2,7 +2,7 @@ normalize_priors <- function(x,
                              fn = function(x){x/sum(x,na.rm = TRUE)},
                              verbose=TRUE){
     if(verbose){
-        print("Normalizing priors with the function:")
+        messager("  Normalizing priors with the function:")
         print(fn)
     } 
     fn(x)