+++ NEWS.md
@@ -1,3 +1,15 @@
+# echoverseTemplate 0.99.1
+
+## New features
+
+* Updated conda env to *echoR_mini*. 
+* Add `prior_k` as argument to `FINEMAP`.
+* Add message about gcc8 to FINEMAP v1.4 error.
+
+## Bug fixes
+
+* Minor fixes to POLYFUN method checks. 
+
 # echoverseTemplate 0.99.0
 
 * Added a `NEWS.md` file to track changes to the package.